Renamed namespace app.blueprints.rest.v2.case_objects into app.blueprints.rest.v2.case_routes
