+++ source/app/blueprints/rest/v2/case_routes/assets.py
