+++ source/app/blueprints/rest/v2/case_routes/notes_directories.py
