+++ source/app/blueprints/rest/v2/cases.py
@@ -29,13 +29,13 @@ from app.blueprints.rest.endpoints import response_api_created
 from app.blueprints.rest.endpoints import response_api_error
 from app.blueprints.rest.endpoints import response_api_paginated
 from app.blueprints.rest.parsing import parse_pagination_parameters
-from app.blueprints.rest.v2.case_objects.assets import case_assets_blueprint
-from app.blueprints.rest.v2.case_objects.iocs import case_iocs_blueprint
-from app.blueprints.rest.v2.case_objects.notes import case_notes_blueprint
-from app.blueprints.rest.v2.case_objects.notes_directories import case_notes_directories_blueprint
-from app.blueprints.rest.v2.case_objects.tasks import case_tasks_blueprint
-from app.blueprints.rest.v2.case_objects.evidences import case_evidences_blueprint
-from app.blueprints.rest.v2.case_objects.events import case_events_blueprint
+from app.blueprints.rest.v2.case_routes.assets import case_assets_blueprint
+from app.blueprints.rest.v2.case_routes.iocs import case_iocs_blueprint
+from app.blueprints.rest.v2.case_routes.notes import case_notes_blueprint
+from app.blueprints.rest.v2.case_routes.notes_directories import case_notes_directories_blueprint
+from app.blueprints.rest.v2.case_routes.tasks import case_tasks_blueprint
+from app.blueprints.rest.v2.case_routes.evidences import case_evidences_blueprint
+from app.blueprints.rest.v2.case_routes.events import case_events_blueprint
 from app.iris_engine.access_control.iris_user import iris_current_user
 from app.business.cases import cases_create
 from app.business.cases import cases_delete
